fix: switch to long int

+++ src/lexer.h
@@ -22,7 +22,7 @@ typedef enum {
 typedef struct token {
     Tok_t token;
     union {
-        int i;
+        long int i;
         double d;
         char *id;
     } value;
